+++ src/app/(dashboard)/media/actions/update-media.action.ts
@@ -6,18 +6,22 @@ import { createClient } from '@/lib/supabase/server';
 import { getUserWithRole } from '@/lib/supabase/auth';
 import { createLogEvent } from '@/lib/supabase/log-event';
 
-import type { MediaEditForm } from '../schemas/edit-form.schema';
+import type { MediaUpdateMediaAction } from '../schemas/actions/update-media.schema';
+import { MediaUpdateMediaActionSchema } from '../schemas/actions/update-media.schema';
 
-interface UpdateMediaParams {
-  mediaId: string;
-  formData: MediaEditForm;
-}
-
-export async function updateMedia({ mediaId, formData }: UpdateMediaParams) {
+export async function updateMedia({ mediaId, formData }: MediaUpdateMediaAction) {
   const { user, role } = await getUserWithRole();
 
   if (!user || !role || !['admin', 'editor'].includes(role)) {
     await createLogEvent('error', 'UPDATE_MEDIA_UNAUTHORIZED', 'Unauthorized user. Media Id:' + mediaId, user?.id);
+
+    return { success: false };
+  }
+
+  const parsed = MediaUpdateMediaActionSchema.safeParse({ mediaId, formData });
+
+  if (!parsed.success) {
+    await createLogEvent('error', 'UPDATE_MEDIA_INVALID_DATA', 'Update media failed schema validation', user?.id);
 
     return { success: false };
   }